chore: remote generation if logged into cloud
