+++ src/redteam/constants.ts
@@ -1,9 +1,3 @@
-import { getEnvString } from '../envars';
-
-export function getRemoteGenerationUrl(): string {
-  return getEnvString('PROMPTFOO_REMOTE_GENERATION_URL', 'https://api.promptfoo.dev/v1/generate');
-}
-
 export const DEFAULT_NUM_TESTS_PER_PLUGIN = 5;
 
 export const REDTEAM_MODEL = 'openai:chat:gpt-4o';
