+++ src/redteam/extraction/purpose.ts
@@ -1,7 +1,7 @@
 import dedent from 'dedent';
 import logger from '../../logger';
 import type { ApiProvider } from '../../types';
-import { shouldGenerateRemote } from '../util';
+import { shouldGenerateRemote } from '../remoteGeneration';
 import type { RedTeamTask } from './util';
 import { callExtraction, fetchRemoteGeneration, formatPrompts } from './util';
 
@@ -21,11 +21,11 @@ export async function extractSystemPurpose(
   // Fallback to local extraction
   const prompt = dedent`
     The following are prompts that are being used to test an LLM application:
-    
+
     ${formatPrompts(prompts)}
-    
+
     Given the above prompts, output the "system purpose" of the application in a single sentence, enclosed in <Purpose> tags.
-    
+
     Example outputs:
     <Purpose>Provide users a way to manage finances</Purpose>
     <Purpose>Executive assistant that helps with scheduling and reminders</Purpose>
