+++ src/remoteGrading.ts
@@ -1,7 +1,7 @@
 import { fetchWithCache } from './cache';
 import logger from './logger';
 import { REQUEST_TIMEOUT_MS } from './providers/shared';
-import { getRemoteGenerationUrl } from './redteam/constants';
+import { getRemoteGenerationUrl } from './redteam/remoteGeneration';
 import type { GradingResult } from './types';
 
 type RemoteGradingPayload = {
