+++ test/redteam/util.test.ts
@@ -1,7 +1,4 @@
-import cliState from '../../src/cliState';
-import { getEnvBool, getEnvString } from '../../src/envars';
-import { neverGenerateRemote, removePrefix } from '../../src/redteam/util';
-import { shouldGenerateRemote } from '../../src/redteam/util';
+import { removePrefix } from '../../src/redteam/util';
 
 describe('removePrefix', () => {
   it('should remove a simple prefix', () => {
@@ -30,64 +27,5 @@ describe('removePrefix', () => {
 
   it('should handle prefix that is the entire string', () => {
     expect(removePrefix('Prompt:', 'Prompt')).toBe('');
-  });
-});
-
-jest.mock('../../src/envars');
-jest.mock('../../src/cliState', () => ({
-  remote: undefined,
-}));
-
-describe('shouldGenerateRemote', () => {
-  beforeEach(() => {
-    jest.resetAllMocks();
-    cliState.remote = undefined;
-  });
-
-  it('should return true when remote generation is not disabled and no OpenAI key exists', () => {
-    jest.mocked(getEnvBool).mockReturnValue(false);
-    jest.mocked(getEnvString).mockReturnValue('');
-    expect(shouldGenerateRemote()).toBe(true);
-  });
-
-  it('should return false when remote generation is disabled via env var', () => {
-    jest.mocked(getEnvBool).mockReturnValue(true);
-    jest.mocked(getEnvString).mockReturnValue('');
-    expect(shouldGenerateRemote()).toBe(false);
-  });
-
-  it('should return false when OpenAI key exists', () => {
-    jest.mocked(getEnvBool).mockReturnValue(false);
-    jest.mocked(getEnvString).mockReturnValue('sk-123');
-    expect(shouldGenerateRemote()).toBe(false);
-  });
-
-  it('should return false when remote generation is disabled via env var and OpenAI key exists', () => {
-    jest.mocked(getEnvBool).mockReturnValue(true);
-    jest.mocked(getEnvString).mockReturnValue('sk-123');
-    expect(shouldGenerateRemote()).toBe(false);
-  });
-
-  it('should return true when cliState.remote is true regardless of other conditions', () => {
-    jest.mocked(getEnvBool).mockReturnValue(true);
-    jest.mocked(getEnvString).mockReturnValue('sk-123');
-    cliState.remote = true;
-    expect(shouldGenerateRemote()).toBe(true);
-  });
-});
-
-describe('neverGenerateRemote', () => {
-  beforeEach(() => {
-    jest.resetAllMocks();
-  });
-
-  it('should return true when PROMPTFOO_DISABLE_REDTEAM_REMOTE_GENERATION is set to true', () => {
-    jest.mocked(getEnvBool).mockReturnValue(true);
-    expect(neverGenerateRemote()).toBe(true);
-  });
-
-  it('should return false when PROMPTFOO_DISABLE_REDTEAM_REMOTE_GENERATION is set to false', () => {
-    jest.mocked(getEnvBool).mockReturnValue(false);
-    expect(neverGenerateRemote()).toBe(false);
   });
 });
